+++ packages/core-js-compat/src/data.mjs
@@ -274,6 +274,14 @@ export const data = {
     safari: '9.0',
     rhino: '1.7.13',
   },
+  'es.array.push': {
+    // bug with setting length
+    // chrome: '103',
+    // edge: '15',
+    firefox: '55',
+    // the same to Chrome bug fixed only in Safari TP
+    // safari: '15.4',
+  },
   'es.array.reduce': {
     chrome: '83', // https://bugs.chromium.org/p/chromium/issues/detail?id=1049982
     firefox: '4',
@@ -338,6 +346,13 @@ export const data = {
     chrome: '73',
     firefox: '67',
     safari: '13',
+  },
+  'es.array.unshift': {
+    chrome: '71',
+    firefox: '23',
+    ie: '9',
+    // bug with setting length fixed only in Safari TP
+    // safari: '15.4',
   },
   'es.array-buffer.constructor': {
     chrome: '26',