+++ packages/core-js-compat/src/modules-by-versions.mjs
@@ -142,6 +142,8 @@ export default {
   3.23: [
     'es.array.find-last',
     'es.array.find-last-index',
+    'es.array.push',
+    'es.array.unshift',
     'es.typed-array.find-last',
     'es.typed-array.find-last-index',
     'esnext.symbol.metadata-key',