+++ packages/core-js/actual/array/unshift.js
@@ -1,0 +1,3 @@
+var parent = require('../../stable/array/unshift');
+
+module.exports = parent;