+++ packages/core-js/es/array/virtual/index.js
@@ -17,6 +17,7 @@ require('../../../modules/es.array.iterator');
 require('../../../modules/es.array.join');
 require('../../../modules/es.array.last-index-of');
 require('../../../modules/es.array.map');
+require('../../../modules/es.array.push');
 require('../../../modules/es.array.reduce');
 require('../../../modules/es.array.reduce-right');
 require('../../../modules/es.array.reverse');
@@ -27,6 +28,7 @@ require('../../../modules/es.array.species');
 require('../../../modules/es.array.splice');
 require('../../../modules/es.array.unscopables.flat');
 require('../../../modules/es.array.unscopables.flat-map');
+require('../../../modules/es.array.unshift');
 require('../../../modules/es.object.to-string');
 var entryVirtual = require('../../../internals/entry-virtual');