+++ packages/core-js/full/array/unshift.js
@@ -1,0 +1,3 @@
+var parent = require('../../actual/array/unshift');
+
+module.exports = parent;