+++ packages/core-js/stable/array/push.js
@@ -1,0 +1,3 @@
+var parent = require('../../es/array/push');
+
+module.exports = parent;